add localization for Root folder label
