+++ core-web/libs/portlets/dot-content-drive/ui/src/lib/dot-tree-folder/dot-tree-folder.component.html
@@ -20,7 +20,7 @@
                 data-testid="tree-node-label"
                 [attr.data-json-node]="node.data | json"
                 [class.active]="$activeDropNode()?.id === node?.data.id">
-                {{ node.label | folderName }}
+                {{ node.key === ALL_FOLDER_KEY ? (node.label | dm) : (node.label | folderName) }}
             </span>
         </ng-template>
         <!-- LOADING WONT WORK WITH CUSTOM TEMPLATE: https://github.com/primefaces/primeng/blob/v17/src/app/components/tree/tree.ts#L113 -->
