+++ dotCMS/src/main/webapp/WEB-INF/messages/Language.properties
@@ -6129,6 +6129,8 @@ content-drive.toast.unlock-success-detail=You’ve unlocked this contentlet for 
 content-drive.toast.unlock-error=Couldn't unlock
 content-drive.toast.unlock-error-detail=Something went wrong. The contentlet wasn’t unlocked.
 
+content-drive.all-folder.label=All
+
 content-drive.toast.download-success=Downloading {0}
 content-drive.toast.download-success-detail=The download has started.
